requirements.txt
braintree==3.10.0
Django==1.9.2
django-cors-headers
django-crispy-forms==1.5.2
django-filter
django-xadmin==0.5.0
djangorestframework
djangorestframework-jwt
httpie
Markdown
markdown2==2.3.0
MarkupSafe
oauthlib
Pillow==3.0.0
Pygments==2.0.2
PyJWT
python-openid
python-social-auth
python3-openid
requests==2.8.1
requests-oauthlib
six==1.10.0
wheel==0.29.0
WTForms==2.0.2
dj-database-url==0.4.0
gunicorn==19.4.5
psycopg2==2.6.1
whitenoise==2.0.6